styling$css refactor
