+++ src/components/Footer/footer.module.scss
@@ -12,14 +12,5 @@
 }
 
 .content {
-  color: #FFFFFF;
-  display: flex;
-  justify-content: space-between;
-  padding: 20px 130px 30px 130px;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 142%;
-  text-align: center;
   background: #1C3B55;
 }
