+++ src/components/Header/header.module.scss
@@ -2,6 +2,5 @@
     background: linear-gradient(90.95deg, #255D8B 2.29%, #3A80BA 96.79%);
 }
 .content{
-
     padding: 24px 0 10px 136px;
 }
