+++ src/components/MainPage/LowerContent/infoItem.module.scss
@@ -12,6 +12,7 @@
     font-size: 24px;
     line-height: 29px;
     text-align: center;
+    color: #1A1A1A;
 }
 .text{
     font-style: normal;
