+++ src/components/MainPage/LowerContent/lowerContent.module.scss
@@ -6,5 +6,4 @@
 .subscribeContent {
     text-align: center;
     padding-top:154px;
-
 }
