+++ src/components/MainPage/UpperContent/upperContent.module.scss
@@ -1,42 +1,33 @@
-/*@font-face {
-  font-family: Pompadur;
-  src: url(src/assets/fonts)
-}*/
 
-
-.textBanner{
-  max-width: 500px;
+.textBanner {
+  max-width: 590px;
 }
-.text{
-  font-family: Montserrat;
+
+.text {
   font-style: normal;
-  font-weight: bold;
+  font-weight: normal;
   font-size: 48px;
   line-height: 59px;
   color: #FFFFFF;
+  margin-bottom: 24px;
 }
-.smallText{
-  font-family: Montserrat;
+
+.lowerText {
   font-style: normal;
   font-weight: normal;
   font-size: 20px;
   line-height: 28px;
   color: #FFFFFF;
 }
-.lowerText{
-  font-weight: 400;
-  font-style: normal;
-  font-size: 20px;
-  color: #FFFFFF;
 
-}
-.button{
+.button {
   width: 154px;
   height: 56px;
-  left: 130px;
-  top: 351px;
   background: #FFFFFF;
   border-radius: 8px;
-  border-color: #FFFFFF;
-  margin-top: 30px;
+  margin-top:32px;
+}
+
+.boldText {
+  font-weight: bold;
 }
