+++ src/components/MainPage/mainPage.module.scss
