+++ src/components/NavigationComponent/NavigationComponent.js
@@ -1,7 +1,7 @@
 import React from "react"
 import {connect} from "react-redux"
 import {NavLink, withRouter} from "react-router-dom"
-import S from "./navigation.module.css"
+import S from "./navigation.module.scss"
 import {compose} from "redux"
 import {clearStore} from "redux/statistic-reducer"
 
@@ -13,12 +13,12 @@ const NavigationComponent = ({navItems, navLinks, ...props}) => {
         let items = navItems.filter(e => e !== '')
         items = items.slice(0, items.length - 1)
         items = items.map((i, n) => {
-            return <span key={n} onClick={(n === 0) ? onClick : null}><NavLink className = {S.item} key={n}
-                to={navLinks[n]}><span key={n}>{i}</span></NavLink>
-                <span className = {S.item}>{">"}</span></span>
+            return <span key={n} onClick={(n === 0) ? onClick : null}><NavLink className={S.item} key={n}
+                                                                               to={navLinks[n]}><span key={n}>{i}</span></NavLink>
+                <span className={S.item}>{">"}</span></span>
         })
         items.push(<span key={"key"}
-            className={S.current}><span>{navItems[items.length]}</span></span>
+                         className={`${S.item} ${S.current}`}><span>{navItems[items.length]}</span></span>
         )
         return items
     }
