+++ src/components/NavigationComponent/navigation.module.scss
@@ -13,9 +13,5 @@
 }
 
 .current {
-    font-style: normal;
-    font-weight: normal;
-    font-size: 16px;
-    line-height: 20px;
     color: #3A80BA;
 }
